requirements.txt
asyncpg==0.29.0
cryptography==43.0.0
fastapi==0.111.1
passlib==1.7.4
psycopg2_binary==2.9.9
PyJWT==2.8.0
SQLAlchemy==2.0.31
uvicorn==0.30.3